Create a DCAT-AP CH Catalog (dcat:Catalog)

Create a DCAT-AP CH Catalog (dcat:Catalog)

Usage

dcat_catalog(
  catalog_endpoint = "https://swisstopo/catalog-endpoint.rdf",
  datasets = "<dcat:dataset rdf:resource=\"https://swisstopo/123\"/>",
  ...
)

Arguments

catalog_endpoint

URL/URI of the catalog endpoint

datasets

character vector with one or more dcat:dataset(s), see dcat_dataset()

...

any other property to be included to a dcat:Catalog

Value

A character vector of length greater than one, where each element is a line of the catalog in rdf format. The catalog can be exported in a file with unix file endings and UTF-8 encoding with write_unix().

Details

A dcat:Catalog class, created with dcat_catalog(), can/must have the following properties (M stands for mandatory; R for recommended):

dcat:dataset, M in V1, R in V2: see dcat_dataset()

dct:description, M in V2: see dct_description()

dct:publisher, M in V2: see dct_publisher()

dct:title, M in V2: see dct_title()

foaf:homepage, R in V2: see foaf_homepage()

dct:language, R in V2: see dct_language()

dct:license, R in V2: see dct_license()

dct:issued, R in V2: see dct_issued()

dct:rights, R in V2: see dct_rights()

dct:spatial, R in V2: see dct_spatial()

dcat:themeTaxonomy, R in V2: see dcat_themeTaxonomy()

dct:modified, R in V2: see dct_modified()

dcat:service, R in V2: see dcat_service()

Examples

dcat_catalog()
#>  [1] "<?xml version=\"1.0\" encoding=\"utf-8\" ?>"                          
#>  [2] "<rdf:RDF "                                                            
#>  [3] "  xmlns:foaf=\"http://xmlns.com/foaf/0.1/\""                          
#>  [4] "  xmlns:rdfs=\"http://www.w3.org/2000/01/rdf-schema#\""               
#>  [5] "  xmlns:hydra=\"http://www.w3.org/ns/hydra/core#\""                   
#>  [6] "  xmlns:rdf=\"http://www.w3.org/1999/02/22-rdf-syntax-ns#\""          
#>  [7] "  xmlns:dcat=\"http://www.w3.org/ns/dcat#\""                          
#>  [8] "  xmlns:dct=\"http://purl.org/dc/terms/\""                            
#>  [9] "  xmlns:vcard=\"http://www.w3.org/2006/vcard/ns#\""                   
#> [10] "  xmlns:xsd=\"http://www.w3.org/2001/XMLSchema#\""                    
#> [11] "  xmlns:schema=\"http://schema.org/\""                                
#> [12] ">"                                                                    
#> [13] "  <dcat:Catalog rdf:about=\"https://swisstopo/catalog-endpoint.rdf\">"
#> [14] "    <dcat:dataset rdf:resource=\"https://swisstopo/123\"/>"           
#> [15] "  </dcat:Catalog>"                                                    
#> [16] "</rdf:RDF>"                                                           
dcat_catalog(catalog_endpoint = "")
#>  [1] "<?xml version=\"1.0\" encoding=\"utf-8\" ?>"                
#>  [2] "<rdf:RDF "                                                  
#>  [3] "  xmlns:foaf=\"http://xmlns.com/foaf/0.1/\""                
#>  [4] "  xmlns:rdfs=\"http://www.w3.org/2000/01/rdf-schema#\""     
#>  [5] "  xmlns:hydra=\"http://www.w3.org/ns/hydra/core#\""         
#>  [6] "  xmlns:rdf=\"http://www.w3.org/1999/02/22-rdf-syntax-ns#\""
#>  [7] "  xmlns:dcat=\"http://www.w3.org/ns/dcat#\""                
#>  [8] "  xmlns:dct=\"http://purl.org/dc/terms/\""                  
#>  [9] "  xmlns:vcard=\"http://www.w3.org/2006/vcard/ns#\""         
#> [10] "  xmlns:xsd=\"http://www.w3.org/2001/XMLSchema#\""          
#> [11] "  xmlns:schema=\"http://schema.org/\""                      
#> [12] ">"                                                          
#> [13] "  <dcat:Catalog>"                                           
#> [14] "    <dcat:dataset rdf:resource=\"https://swisstopo/123\"/>" 
#> [15] "  </dcat:Catalog>"                                          
#> [16] "</rdf:RDF>"                                                 

# Add some properties (...)
dcat_catalog(
  title = dct_title("Catalogo Ustat", language = "it"),
  description = dct_description("Descrizione catalogo", language = "it")
)
#>  [1] "<?xml version=\"1.0\" encoding=\"utf-8\" ?>"                                
#>  [2] "<rdf:RDF "                                                                  
#>  [3] "  xmlns:foaf=\"http://xmlns.com/foaf/0.1/\""                                
#>  [4] "  xmlns:rdfs=\"http://www.w3.org/2000/01/rdf-schema#\""                     
#>  [5] "  xmlns:hydra=\"http://www.w3.org/ns/hydra/core#\""                         
#>  [6] "  xmlns:rdf=\"http://www.w3.org/1999/02/22-rdf-syntax-ns#\""                
#>  [7] "  xmlns:dcat=\"http://www.w3.org/ns/dcat#\""                                
#>  [8] "  xmlns:dct=\"http://purl.org/dc/terms/\""                                  
#>  [9] "  xmlns:vcard=\"http://www.w3.org/2006/vcard/ns#\""                         
#> [10] "  xmlns:xsd=\"http://www.w3.org/2001/XMLSchema#\""                          
#> [11] "  xmlns:schema=\"http://schema.org/\""                                      
#> [12] ">"                                                                          
#> [13] "  <dcat:Catalog rdf:about=\"https://swisstopo/catalog-endpoint.rdf\">"      
#> [14] "    <dct:title xml:lang=\"it\">Catalogo Ustat</dct:title>"                  
#> [15] "    <dct:description xml:lang=\"it\">Descrizione catalogo</dct:description>"
#> [16] "    <dcat:dataset rdf:resource=\"https://swisstopo/123\"/>"                 
#> [17] "  </dcat:Catalog>"                                                          
#> [18] "</rdf:RDF>"